creaded a dedicated vector db
The api design is still bad. the main db which remembers past embeddings
is different from the index, on which we query. Index has everything it
needs to generate the output.

The design is such because index creation is costly, one time job, with
heavy system requirements. Index querying is itself, quite simple. There
is not lot to it. And this needs to be quite lightweight.